backup/requirements.txt
pandas
numpy
streamlit
acoustics
librosa
openpyxl
pyfunctional
scipy